+++ examples/demo3/src/index/App.vue
@@ -3,12 +3,20 @@
     <div class="group" v-for="item in list" :key="item">
       <div class="label">{{item}}</div>
       <div class="comp">
-        <view v-if="item === 'view'">我是视图</view>
+        <!-- 可使用 html 标签替代的内置组件 -->
         <img v-if="item === 'image'" src="https://res.wx.qq.com/wxdoc/dist/assets/img/0.4cb08bb4.jpg" width="50" height="50" />
         <input v-else-if="item === 'input'" type="number" placeholder="请输入内容" @input="onInput" />
         <textarea v-else-if="item === 'textarea'" placeholder="请输入内容" maxlength="50" :auto-height="true" value="我是 textarea" @input="onTextareaInput" />
-        <wx-component v-else-if="item === 'picker'" behavior="picker" mode="region" @change="onPickerChange">选择城市</wx-component>
         <video v-else-if="item === 'video'" src="http://wxsnsdy.tc.qq.com/105/20210/snsdyvideodownload?filekey=30280201010421301f0201690402534804102ca905ce620b1241b726bc41dcff44e00204012882540400&bizid=1023&hy=SH&fileparam=302c020101042530230204136ffd93020457e3c4ff02024ef202031e8d7f02030f42400204045a320a0201000400" :muted="true" :show-mute-btn="true" :controls="true"></video>
+        <!-- 使用 wx-component 来创建内置组件 -->
+        <wx-component v-else-if="item === 'view'" behavior="view">我是视图</wx-component>
+        <wx-component v-else-if="item === 'button'" behavior="button" open-type="share">分享</wx-component>
+        <wx-component v-else-if="item === 'picker'" behavior="picker" mode="region" @change="onPickerChange">
+          <span>点击&nbsp;&nbsp;</span>
+          <span>选择城市</span>
+        </wx-component>
+        <!-- 不支持标签 -->
+        <iframe v-else-if="item === 'iframe'"></iframe>
       </div>
     </div>
   </div>
@@ -21,13 +29,20 @@ export default {
   data() {
     return {
       list: [
-        'view',
         'image',
         'input',
         'textarea',
-        'picker',
         'video',
+        'view',
+        'button',
+        'picker',
+        'iframe',
       ],
+    }
+  },
+  created() {
+    window.onDealWithNotSupportDom = dom => {
+      dom.textContent = `标签 ${dom.tagName.toLowerCase()} 暂不支持`
     }
   },
   methods: {