+++ packages/miniprogram-element/src/component/video.js
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
 const mp = require('miniprogram-render')
 
 const {