+++ packages/miniprogram-element/src/util/init-handle.js
@@ -1,38 +1,42 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
 const mp = require('miniprogram-render')
 const _ = require('./tool')
+const view = require('../component/view')
 const video = require('../component/video')
 const image = require('../component/image')
 const input = require('../component/input')
 const textarea = require('../component/textarea')
 const picker = require('../component/picker')
+const button = require('../component/button')
 
 const {
     Event,
 } = mp.$$adapter
-const WX_COMP_NAME_MAP = {
-    picker: 'picker',
-    IMG: 'image',
-    INPUT: 'input',
-    TEXTAREA: 'textarea',
-    VIDEO: 'video',
-}
+const {
+    WX_COMP_NAME_MAP,
+    NOT_SUPPORT,
+} = _
 
 module.exports = {
-    WX_COMP_NAME_MAP,
-
     /**
      * 初始化
      */
     init(data) {
-        const tagName = this.domNode.tagName
+        const domNode = this.domNode
+        const tagName = domNode.tagName
+
         if (tagName === 'WX-COMPONENT') {
-            this.initWxComponent(data)
-        } else if (tagName === 'IFRAME') {
-            this.initNotSupport(data) // 因为无法支持 iframe，所以需要显示提示文字
+            // 无可替换 html 标签
+            data.wxCompName = domNode.$$behavior
+            const wxCompName = WX_COMP_NAME_MAP[data.wxCompName]
+            if (wxCompName) _.checkComponentAttr(wxCompName, domNode, data)
+        } else if (NOT_SUPPORT.indexOf(tagName) >= 0) {
+            // 不支持标签
+            data.wxCompName = 'not-support'
+            data.content = domNode.textContent
         } else {
+            // 可替换 html 标签
             const wxCompName = WX_COMP_NAME_MAP[tagName]
-            if (wxCompName) _.checkComponentAttr(wxCompName, this.domNode, data)
+            if (wxCompName) _.checkComponentAttr(wxCompName, domNode, data)
         }
     },
 
@@ -53,36 +57,11 @@ module.exports = {
         })
     },
 
-    /**
-     * 特殊内置组件
-     */
-    initWxComponent(data) {
-        data.wxCompName = this.domNode.$$behavior
-        data.content = this.domNode.$$content
-        data.class = this.domNode.$$domInfo.class || ''
-        data.style = this.domNode.style.cssText
-
-        if (this.domNode.$$behavior === 'button') {
-            _.checkAttrUpdate(this.data, this.domNode, data, ['disabled', 'openType'])
-        } else {
-            const wxCompName = WX_COMP_NAME_MAP[this.domNode.$$behavior]
-            if (wxCompName) _.checkComponentAttr(wxCompName, this.domNode, data)
-        }
-    },
-
-    /**
-     * 不支持组件，均转换成 view
-     */
-    initNotSupport(data) {
-        _.checkComponentAttr('view', this.domNode, data)
-
-        // TODO: remove
-        data.content = this.domNode.$$content
-    },
-
+    ...view.handles,
     ...input.handles,
     ...textarea.handles,
     ...video.handles,
     ...picker.handles,
     ...image.handles,
+    ...button.handles,
 }