+++ packages/miniprogram-element/test/utils.js
@@ -1,6 +1,6 @@
 const path = require('path')
-// eslint-disable-next-line import/no-extraneous-dependencies
 const mp = require('miniprogram-render')
+// eslint-disable-next-line import/no-extraneous-dependencies
 const simulate = require('miniprogram-simulate')
 
 /**